chore(litmus-sdk): Update the README for sdk cli
